+++ src/Psl/File/read.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Psl\Filesystem;
+namespace Psl\File;
 
 use Psl;
-use Psl\File;
 use Psl\IO;
 use Psl\Str;
 
@@ -21,11 +20,11 @@ use Psl\Str;
  *                                                   $file does not exist, or is not readable.
  * @throws Exception\RuntimeException In case of an error.
  */
-function read_file(string $file, int $offset = 0, ?int $length = null): string
+function read(string $file, int $offset = 0, ?int $length = null): string
 {
     try {
-        $handle = File\open_read_only($file);
-        $lock = $handle->lock(File\LockType::SHARED);
+        $handle = namespace\open_read_only($file);
+        $lock = $handle->lock(namespace\LockType::SHARED);
 
         $handle->seek($offset);
         $content = $handle->readAll($length);
@@ -34,7 +33,7 @@ function read_file(string $file, int $offset = 0, ?int $length = null): string
         $handle->close();
 
         return $content;
-    } catch (File\Exception\ExceptionInterface | IO\Exception\ExceptionInterface $previous) {
+    } catch (IO\Exception\ExceptionInterface $previous) {
         // @codeCoverageIgnoreStart
         throw new Exception\RuntimeException(Str\format(
             'Failed to read file "%s".',