+++ src/Psl/File/write.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Psl\Filesystem;
+namespace Psl\File;
 
 use Psl;
 use Psl\File;
+use Psl\Filesystem;
 use Psl\IO;
 use Psl\Str;
 
@@ -14,25 +15,17 @@ use function clearstatcache;
 /**
  * Write $content to $file.
  *
- * If $file does not exist, it will be created.
- *
  * @param non-empty-string $file
  *
  * @throws Psl\Exception\InvariantViolationException If the file specified by
  *                                                   $file is a directory, or is not writeable.
  * @throws Exception\RuntimeException In case of an error.
  */
-function write_file(string $file, string $content): void
+function write(string $file, string $content, WriteMode $mode = WriteMode::OPEN_OR_CREATE): void
 {
     clearstatcache();
 
     try {
-        if (namespace\is_file($file)) {
-            $mode = File\WriteMode::TRUNCATE;
-        } else {
-            $mode = File\WriteMode::OPEN_OR_CREATE;
-        }
-
         $handle = File\open_write_only($file, $mode);
         $lock = $handle->lock(File\LockType::EXCLUSIVE);
 
@@ -42,7 +35,7 @@ function write_file(string $file, string $content): void
         $handle->close();
 
         clearstatcache();
-    } catch (File\Exception\ExceptionInterface | IO\Exception\ExceptionInterface $previous) {
+    } catch (IO\Exception\ExceptionInterface | Filesystem\Exception\ExceptionInterface $previous) {
         // @codeCoverageIgnoreStart
         throw new Exception\RuntimeException(Str\format(
             'Failed to write to file "%s".',